Make test_reforms fixture have session scope
